+++ releasenotes.md
@@ -46,6 +46,7 @@
 - Type function `tagof` is renamed `get_tag`.
 - Add `untypedlist` as a usable type #2647.
 - `??` and `?:` has new precedence and binds tighter than `+` and `-`
+- Added the `tags` property for types and `$reflect`.
 
 ### Stdlib changes
 - `std::collections::RingBuffer` has been renamed `RingList`.